# RFC: V3: Oxc resolver