--- a/crates/atlaspack_plugin_resolver_oxc/Cargo.toml
+++ b/crates/atlaspack_plugin_resolver_oxc/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "atlaspack_plugin_resolver_oxc"
+version = "0.1.0"
+edition = "2021"
+description = "Resolver Plugin for the Atlaspack Bundler"
+
+[dependencies]
+atlaspack_core = { path = "../atlaspack_core" }
+anyhow = { workspace = true }
+serde = { workspace = true }
+async-trait = { workspace = true }
+oxc_resolver = { workspace = true }
+normalize-path = { workspace = true }
+
+[dev-dependencies]
+atlaspack_filesystem = { path = "../atlaspack_filesystem" }