Maintaining our own Nodejs resolver is quite difficult, error prone and increases the maintenance/testing surface of Atlapack. This RFC is about replacing our native resolver with an oxc based resolver and supplies a reference for testing.

Using our benchmarks, oxc is 10% faster (though the actual plugin is doing less so YMMV), would like to test this against JFE to see if it resolves the tsconfig issues we are seeing

To use this just add `@atlaspack/resolver-oxc` to the `.parcelrc` file