@devongovett 2024-11-10
FYI, I made Parcel's resolver a lot faster recently (basically rewrote the whole cache): https://github.com/parcel-bundler/parcel/pull/9984

Benchmarked a bunch of resolver crates including OXC: https://github.com/parcel-bundler/parcel/blob/v2/crates/parcel-resolver/benches/benchmark.rs

With caching:

```
cached/parcel_resolver  time:   [1.8614 µs 1.8661 µs 1.8716 µs]
cached/oxc_resolver     time:   [6.4001 µs 6.4171 µs 6.4362 µs]
cached/rspack_resolver  time:   [6.3623 µs 6.3738 µs 6.3856 µs]
cached/deno             time:   [81.661 µs 82.007 µs 82.467 µs]
```

Without caching (creating a fresh cache on each iteration):

```
uncached/parcel_resolver time:   [109.29 µs 109.70 µs 110.13 µs]
uncached/oxc_resolver    time:   [162.42 µs 163.24 µs 164.28 µs]
uncached/rspack_resolver time:   [165.43 µs 166.10 µs 167.01 µs]
uncached/deno            time:   [157.49 µs 157.66 µs 157.88 µs]
```

Published it on crates.io as well in case you want to use it. https://docs.rs/parcel-resolver/