@IWANABETHATGUY 2024-11-10
Here is the data in Ubuntu 22.04, SSD, cpu 5950x
**uncached**
![image](https://github.com/user-attachments/assets/51947d76-150e-4a93-8584-38e970152db1)
**cached**
![image](https://github.com/user-attachments/assets/a03bbe40-011a-49b1-b10f-9e56364188a3)