@yyx990803 2024-11-10
I ran the benchmark myself and the difference doesn't seem to be this significant:

### M2 Pro Macbook Pro

Uncached
```
uncached/parcel_resolver
                        time:   [41.279 µs 41.466 µs 41.758 µs]
Found 10 outliers among 100 measurements (10.00%)
  3 (3.00%) low mild
  4 (4.00%) high mild
  3 (3.00%) high severe

uncached/oxc_resolver   time:   [51.383 µs 51.610 µs 51.859 µs]
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe

uncached/rspack_resolver
                        time:   [51.416 µs 51.687 µs 51.992 µs]
Found 14 outliers among 100 measurements (14.00%)
  7 (7.00%) high mild
  7 (7.00%) high severe

uncached/deno           time:   [60.985 µs 61.015 µs 61.050 µs]
Found 15 outliers among 100 measurements (15.00%)
  5 (5.00%) low mild
  5 (5.00%) high mild
  5 (5.00%) high severe
```
  
Cached
```
cached/parcel_resolver  time:   [1.7027 µs 1.7062 µs 1.7105 µs]
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe

cached/oxc_resolver     time:   [2.5719 µs 2.5753 µs 2.5803 µs]
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  3 (3.00%) high severe

cached/rspack_resolver  time:   [2.5542 µs 2.5672 µs 2.5908 µs]
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) high mild
  6 (6.00%) high severe

cached/deno             time:   [50.357 µs 50.459 µs 50.591 µs]
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
```

### Linux Ubuntu 22.04 LTS

- CPU: (32) x64 AMD Ryzen 9 5950X 16-Core Processor
- Memory: 64 GB

![resolver-bench1](https://github.com/user-attachments/assets/88924356-3545-4bc5-9be7-5d009c9c96c3)
![resolver-bench](https://github.com/user-attachments/assets/c2406eb3-73f3-475f-8a82-df36c118be59)