@yamadapc 2024-11-10
I think it does make sense to maintain a resolver because we would rather have no resolver plugins.

Instead, we'd add the few patches we are currently using on resolver plugins to the main rust code.

- - -

@devongovett we can't use parcel resolver because it depends on `dashmap` which we have now validated was causing segmentation faults on our developers machines. There were hundreds of native crashes every day, which are completely fixed by https://github.com/atlassian-labs/atlaspack/pull/10.

That is, there is some unsoundness in the `dashmap` crate which leads to memory corruption and then crashes under a certain workload, which we started to hit. Developers were unable to start the server and make a few builds without a crash prior to this change.

- - -

Regarding benchmark, while it is good to micro-benchmark, it's better to measure a real workload.

We have an internal benchmark, which replays all imports in our application.

On that benchmark, the current master branch of this repository performs as well as the `parcel` master branch before cache optimisation ; the cache optimisation changes made `parcel` resolution slower by a non-negligible amount (10%+).

Most importantly, the absolute times to resolve all imports in the repository are very low when they are executed concurrently (a few seconds to resolve all the imports with multi-threading).

I think that if resolution is a build time bottle-neck today, it is only because the builds are based on crawling the module tree, and it'll be much more efficient to crawl the directory tree directly to parse and discover imports, then run resolution as a second time.

- - -

Note that despite around 30% slower resolver (at the moment) in the master branch, and previously we had a 3x slower resolver than parcel master https://github.com/atlassian-labs/atlaspack/pull/156, the actual build times for this repository are 2x faster to build jira than what is in mainline parcel due to a few optimisations on packaging/optimisation phases.

There is currently some degradation in perf. of the master branch **resolver**, when running new experimental code for building the asset graph in rust. This might be related to removal of `Cow` on certain resolver parts. I am unconvinced it is a performance bottle-neck, but there are other fixes to it.