@devongovett 2024-11-10
FWIW, oxc_resolver also uses dashmap. Dashmap v6 was recently released as well. Maybe that fixed some issues you saw.