@yamadapc 2024-11-10
I was hoping that was the case too, but I had looked at the diff between those dashmap versions and there did not seem to be significant changes that could fix the crashes.

What we have now (thread local caches) performs really close to the concurrent dashmap but has much less surface area for unsoundness and is easier to understand. 

It'd be nice to review this at some point but we were having serious unreliability issues related to crashes (they were common, and certain engineers would experience crashes on every single run of certain builds).