@alshdavid 2024-11-14
Will continue this discussion after I get back